tfcbackend/requirements.txt
# Python libraries required for this project.
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0

# Installed as a result of installing Django.
asgiref==3.5.2
sqlparse==0.4.3

# Installed as a result of installing Django REST Framework.
pytz==2022.6

# Installed as a result of installing DRF Simple JWT
pyjwt==2.6.0

